# backend-core-client/models/Check.ts
// tslint:disable
/**
 * Core
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface Check
 */
export interface Check {
    /**
     * @type {string}
     * @memberof Check
     */
    slug: string;
    /**
     * @type {string}
     * @memberof Check
     */
    name: string;
    /**
     * @type {string}
     * @memberof Check
     */
    description?: string;
    /**
     * @type {string}
     * @memberof Check
     */
    status: CheckStatusEnum;
    /**
     * @type {Array<number>}
     * @memberof Check
     */
    latencies: Array<number>;
}

/**
 * @export
 * @enum {string}
 */
export enum CheckStatusEnum {
    Online = 'ONLINE',
    Degraded = 'DEGRADED',
    Offline = 'OFFLINE'
}