# backend-core-client/models/Group.ts
// tslint:disable
/**
 * Core
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface Group
 */
export interface Group {
    /**
     * @type {string}
     * @memberof Group
     */
    slug: string;
    /**
     * @type {string}
     * @memberof Group
     */
    name: string;
    /**
     * @type {string}
     * @memberof Group
     */
    description?: string;
    /**
     * @type {string}
     * @memberof Group
     */
    status: GroupStatusEnum;
}

/**
 * @export
 * @enum {string}
 */
export enum GroupStatusEnum {
    Online = 'ONLINE',
    Degraded = 'DEGRADED',
    Offline = 'OFFLINE'
}