# backend-core-client/models/Project.ts
// tslint:disable
/**
 * Core
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface Project
 */
export interface Project {
    /**
     * @type {string}
     * @memberof Project
     */
    slug: string;
    /**
     * @type {string}
     * @memberof Project
     */
    name: string;
    /**
     * @type {string}
     * @memberof Project
     */
    description?: string;
    /**
     * @type {string}
     * @memberof Project
     */
    status: ProjectStatusEnum;
}

/**
 * @export
 * @enum {string}
 */
export enum ProjectStatusEnum {
    Online = 'ONLINE',
    Degraded = 'DEGRADED',
    Offline = 'OFFLINE'
}