# backend-core/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
		"/status/groups": {
			"get": {
				"operationId": "StatusController_readGroups",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Group" } } } }
					}
				},
				"tags": ["status"]
			}
		},
		"/status/groups/{groupSlug}/projects": {
			"get": {
				"operationId": "StatusController_readProjects",
				"parameters": [{ "name": "groupSlug", "required": true, "in": "path", "schema": { "type": "string" } }],
				"responses": {
					"200": {
						"description": "",
						"content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } }
					}
				},
				"tags": ["status"]
			}
		},
		"/status/groups/{groupSlug}/projects/{projectSlug}/checks": {
			"get": {
				"operationId": "StatusController_readChecks",
				"parameters": [
					{ "name": "groupSlug", "required": true, "in": "path", "schema": { "type": "string" } },
					{ "name": "projectSlug", "required": true, "in": "path", "schema": { "type": "string" } }
				],
				"responses": {
					"200": {
						"description": "",
						"content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Check" } } } }
					}
				},
				"tags": ["status"]
			}
		}
	},
	"info": { "title": "Core", "description": "", "version": "1.0.0", "contact": {} },
	"tags": [],
	"servers": [{ "url": "/api/core", "description": "Current installation" }],
	"components": {
		"securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
		"schemas": {
			"Group": {
				"type": "object",
				"properties": {
					"slug": { "type": "string" },
					"name": { "type": "string" },
					"description": { "type": "string" },
					"status": { "type": "string", "enum": ["ONLINE", "DEGRADED", "OFFLINE"] }
				},
				"required": ["slug", "name", "status"]
			},
			"Project": {
				"type": "object",
				"properties": {
					"slug": { "type": "string" },
					"name": { "type": "string" },
					"description": { "type": "string" },
					"status": { "type": "string", "enum": ["ONLINE", "DEGRADED", "OFFLINE"] }
				},
				"required": ["slug", "name", "status"]
			},
			"Check": {
				"type": "object",
				"properties": {
					"slug": { "type": "string" },
					"name": { "type": "string" },
					"description": { "type": "string" },
					"status": { "type": "string", "enum": ["ONLINE", "DEGRADED", "OFFLINE"] },
					"latencies": { "type": "array", "items": { "type": "number" } }
				},
				"required": ["slug", "name", "status", "latencies"]
			}
		}
	}
}